Added projects tests

Closes DOCUFAMILY-516

See merge request docudots/banking-admin-backend!437

+++ api/app/GraphQL/Mutations/ProjectMutator.php
@@ -35,10 +35,12 @@ class ProjectMutator extends BaseMutator
 
         $project->update($args);
 
-        $project->projectSettings()->delete();
+        if (isset($args['project_settings'])) {
+            $project->projectSettings()->delete();
 
-        foreach ($args['project_settings'] as $settings) {
-            $project->projectSettings()->create($settings);
+            foreach ($args['project_settings'] as $settings) {
+                $project->projectSettings()->create($settings);
+            }
         }
 
         return $project;
